Refactor and Add Docs
