+++ Cargo.toml
@@ -3,6 +3,10 @@ name = "spotuify"
 version = "0.1.0"
 edition = "2024"
 
+[lib]
+name = "spotuify"
+path = "src/lib.rs"
+
 [dependencies]
 open = "5.3.4"
 rspotify = { version = "0.16.1", default-features = false, features = ["client-ureq", "ureq-rustls-tls-native-certs"] }
