+++ src/config.rs
@@ -4,20 +4,38 @@ use std::path::PathBuf;
 
 use serde::Deserialize;
 
+/// XDG config directory prefix used by this application.
 const APP_CONFIG_DIR: &str = "spotuify";
+/// Primary user config file name resolved under the XDG config directory.
 const CONFIG_FILE_NAME: &str = "config.toml";
 
 /// Application configuration loaded from the user's config file.
 #[derive(Debug, Clone, Default, Deserialize, PartialEq, Eq)]
 #[serde(default)]
-pub(crate) struct AppConfig {
+pub struct AppConfig {
     /// Spotify-specific configuration required by the auth flow.
-    pub(crate) spotify: SpotifyConfig,
+    pub spotify: SpotifyConfig,
 }
 
 impl AppConfig {
     /// Parses application configuration from TOML.
-    pub(crate) fn from_toml(input: &str) -> Result<Self, toml::de::Error> {
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use spotuify::config::AppConfig;
+    ///
+    /// let config = AppConfig::from_toml(
+    ///     r#"
+    ///         [spotify]
+    ///         client_id = "client-id"
+    ///         redirect_uri = "http://127.0.0.1:8888/callback"
+    ///     "#,
+    /// ).unwrap();
+    ///
+    /// assert_eq!(config.spotify.client_id.as_deref(), Some("client-id"));
+    /// ```
+    pub fn from_toml(input: &str) -> Result<Self, toml::de::Error> {
         toml::from_str(input)
     }
 
@@ -40,16 +58,16 @@ impl AppConfig {
 /// Spotify settings used to start the authorization flow.
 #[derive(Debug, Clone, Default, Deserialize, PartialEq, Eq)]
 #[serde(default)]
-pub(crate) struct SpotifyConfig {
+pub struct SpotifyConfig {
     /// Spotify application client ID.
-    pub(crate) client_id: Option<String>,
+    pub client_id: Option<String>,
     /// Redirect URI registered for the Spotify application.
-    pub(crate) redirect_uri: Option<String>,
+    pub redirect_uri: Option<String>,
 }
 
 /// Startup config bootstrap state.
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum ConfigBootstrap {
+pub enum ConfigBootstrap {
     /// Config loaded successfully and has the required Spotify settings.
     Ready {
         /// Path used to load the config file.
@@ -68,7 +86,7 @@ pub(crate) enum ConfigBootstrap {
 
 impl ConfigBootstrap {
     /// Loads config from the default XDG config path.
-    pub(crate) fn load() -> Self {
+    pub fn load() -> Self {
         match expected_config_path() {
             Some(path) => Self::load_from_path(path),
             None => Self::NeedsSetup {
@@ -79,7 +97,19 @@ impl ConfigBootstrap {
     }
 
     /// Loads config from a specific file path.
-    pub(crate) fn load_from_path(path: impl Into<PathBuf>) -> Self {
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// use spotuify::config::ConfigBootstrap;
+    ///
+    /// let bootstrap = ConfigBootstrap::load_from_path("/tmp/spotuify/config.toml");
+    /// assert!(matches!(
+    ///     bootstrap,
+    ///     ConfigBootstrap::Ready { .. } | ConfigBootstrap::NeedsSetup { .. }
+    /// ));
+    /// ```
+    pub fn load_from_path(path: impl Into<PathBuf>) -> Self {
         let path = path.into();
         let contents = match fs::read_to_string(&path) {
             Ok(contents) => contents,
@@ -119,12 +149,12 @@ impl ConfigBootstrap {
     }
 
     /// Reports whether startup can proceed directly to auth.
-    pub(crate) fn is_ready(&self) -> bool {
+    pub fn is_ready(&self) -> bool {
         matches!(self, Self::Ready { .. })
     }
 
     /// Builds a concise user-facing config status for the current shell.
-    pub(crate) fn status_label(&self) -> String {
+    pub fn status_label(&self) -> String {
         match self {
             Self::Ready { path, config } => {
                 let redirect_uri = config
@@ -159,7 +189,7 @@ impl Default for ConfigBootstrap {
 
 /// Reason startup cannot use the current config.
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum ConfigIssue {
+pub enum ConfigIssue {
     /// The expected config file does not exist.
     MissingFile,
     /// The expected config file path could not be resolved.
@@ -195,7 +225,7 @@ impl std::fmt::Display for ConfigIssue {
 
 /// Required Spotify config fields.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum RequiredConfigField {
+pub enum RequiredConfigField {
     /// `spotify.client_id`
     SpotifyClientId,
     /// `spotify.redirect_uri`
@@ -218,7 +248,7 @@ fn expected_config_path() -> Option<PathBuf> {
 }
 
 /// Resolves a file path inside the application's XDG config directory.
-pub(crate) fn app_config_file_path(file_name: &str) -> Option<PathBuf> {
+pub fn app_config_file_path(file_name: &str) -> Option<PathBuf> {
     xdg::BaseDirectories::with_prefix(APP_CONFIG_DIR).get_config_file(file_name)
 }
 
@@ -230,15 +260,6 @@ fn is_blank(value: Option<&str>) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    fn temp_config_path(test_name: &str) -> PathBuf {
-        let path = std::env::temp_dir().join(format!(
-            "spotuify-{test_name}-{}-config.toml",
-            std::process::id()
-        ));
-        let _ = fs::remove_file(&path);
-        path
-    }
 
     #[test]
     fn parses_valid_config_toml() {
@@ -257,90 +278,5 @@ mod tests {
             Some("http://127.0.0.1:8888/callback")
         );
         assert!(config.missing_required_fields().is_empty());
-    }
-
-    #[test]
-    fn load_from_path_reports_missing_file() {
-        let path = temp_config_path("missing");
-
-        let bootstrap = ConfigBootstrap::load_from_path(path.clone());
-
-        assert_eq!(
-            bootstrap,
-            ConfigBootstrap::NeedsSetup {
-                path: Some(path),
-                issue: ConfigIssue::MissingFile,
-            }
-        );
-    }
-
-    #[test]
-    fn load_from_path_reports_invalid_toml() {
-        let path = temp_config_path("invalid-toml");
-        fs::write(&path, "not toml =").expect("test config should be writable");
-
-        let bootstrap = ConfigBootstrap::load_from_path(path.clone());
-        let _ = fs::remove_file(&path);
-
-        assert!(matches!(
-            bootstrap,
-            ConfigBootstrap::NeedsSetup {
-                path: Some(_),
-                issue: ConfigIssue::ParseFailed(_),
-            }
-        ));
-    }
-
-    #[test]
-    fn load_from_path_reports_missing_client_id() {
-        let path = temp_config_path("missing-client-id");
-        fs::write(
-            &path,
-            r#"
-                [spotify]
-                redirect_uri = "http://127.0.0.1:8888/callback"
-            "#,
-        )
-        .expect("test config should be writable");
-
-        let bootstrap = ConfigBootstrap::load_from_path(path.clone());
-        let _ = fs::remove_file(&path);
-
-        assert_eq!(
-            bootstrap,
-            ConfigBootstrap::NeedsSetup {
-                path: Some(path),
-                issue: ConfigIssue::MissingRequiredFields(vec![
-                    RequiredConfigField::SpotifyClientId,
-                ]),
-            }
-        );
-    }
-
-    #[test]
-    fn load_from_path_reports_blank_redirect_uri() {
-        let path = temp_config_path("blank-redirect-uri");
-        fs::write(
-            &path,
-            r#"
-                [spotify]
-                client_id = "client-id"
-                redirect_uri = "   "
-            "#,
-        )
-        .expect("test config should be writable");
-
-        let bootstrap = ConfigBootstrap::load_from_path(path.clone());
-        let _ = fs::remove_file(&path);
-
-        assert_eq!(
-            bootstrap,
-            ConfigBootstrap::NeedsSetup {
-                path: Some(path),
-                issue: ConfigIssue::MissingRequiredFields(vec![
-                    RequiredConfigField::SpotifyRedirectUri,
-                ]),
-            }
-        );
     }
 }
